# CbrApp/src/main/java/org/tbank/cbrapp/DTO/CurrencyRateResponse.java
package org.tbank.cbrapp.DTO;

import lombok.AllArgsConstructor;
import lombok.Data;

// Ответ на GET /currencies/rates/{code}
@Data
@AllArgsConstructor
public class CurrencyRateResponse {
    private String currency;
    private double rate;
}